pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factorybench"
version = "0.1.0"
description = "FactoryBench: lean benchmark for industrial troubleshooting"
requires-python = ">=3.11"
readme = "README.md"
authors = [{name="Forgis"}]
dependencies = [
    "click>=8.1.7",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "datasets>=2.20.0",
    "openai>=1.43.0",
    "numpy>=1.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["factorybench*"]
